# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ProPangkat API",
    "version": "4.1.0",
    "description": "Dokumentasi API utama untuk sistem ProPangkat."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login pengguna",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nip", "password"],
                "properties": {
                  "nip": { "type": "string", "example": "198501012010011001" },
                  "password": { "type": "string", "example": "198501012010011001" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "401": { "description": "Kredensial tidak valid" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Success" }
        }
      },
...
    }
  }
}